Make uefi boot entries made by Rescatux smaller.
This avoid a bug that happens when running grub-install . If the boot entry length is too big then the uefi boot entry somehow gets removed.

It might not avoid problems with filenames that are too large but it's better than our first algorithm that added the big '(Rescapp added)' and then the full filename path.